+++ androidApp/src/main/java/com/example/tasky/android/alarm/data/AlarmScheduler.kt
@@ -5,10 +5,10 @@ import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import com.example.tasky.alarm.domain.IAlarmRepository
+import com.example.tasky.alarm.domain.IAlarmScheduler
 import com.example.tasky.alarm.domain.model.AgendaAlarm
-import com.example.tasky.android.alarm.domain.IAlarmScheduler
+import com.example.tasky.alarm.domain.model.NotificationData
 import com.example.tasky.android.alarm.domain.mapper.toNotificationDataParcelable
-import com.example.tasky.android.alarm.domain.model.NotificationData
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toInstant
