+++ androidApp/src/main/java/com/example/tasky/android/alarm/di/AlarmModule.kt
@@ -2,8 +2,8 @@ package com.example.tasky.android.alarm.di
 
 import com.example.tasky.alarm.data.AlarmRepository
 import com.example.tasky.alarm.domain.IAlarmRepository
+import com.example.tasky.alarm.domain.IAlarmScheduler
 import com.example.tasky.android.alarm.data.AlarmScheduler
-import com.example.tasky.android.alarm.domain.IAlarmScheduler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import org.koin.core.module.dsl.singleOf
