+++ androidApp/src/main/java/com/example/tasky/android/alarm/domain/mapper/NotificationDataMapper.kt
@@ -1,7 +1,7 @@
 package com.example.tasky.android.alarm.domain.mapper
 
 import com.example.tasky.agenda.domain.model.AgendaItem
-import com.example.tasky.android.alarm.domain.model.NotificationData
+import com.example.tasky.alarm.domain.model.NotificationData
 import com.example.tasky.common.domain.util.toLocalDateTime
 
 fun AgendaItem.toNotificationData(requestCode: Int? = null) =
