+++ androidApp/src/main/java/com/example/tasky/android/alarm/domain/mapper/NotificationDataParcelableMapper.kt
@@ -1,6 +1,6 @@
 package com.example.tasky.android.alarm.domain.mapper
 
-import com.example.tasky.android.alarm.domain.model.NotificationData
+import com.example.tasky.alarm.domain.model.NotificationData
 import com.example.tasky.android.alarm.domain.model.NotificationDataParcelable
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toInstant
